Fix projection expression in flatten
